CNL-13: File uploading initial draft. Refactoring

+++ src/main/java/com/vansisto/cookandlaughbe/entity/ActivationCode.java
@@ -26,7 +26,7 @@ public class ActivationCode {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long activationCodeId;
+    private Integer activationCodeId;
 
     @Column(unique = true)
     private String code;
